requirements.txt
algoliasearch-django>=2.0,<3.0
django>=4.0.0,<4.1.0
djangorestframework
djangorestframework-simplejwt
requests
pyyaml
django-cors-headers
